unvalidated: parse unvalidated properties
Added functions to parse `Unvalidated$` from an express request. If the
desired property is not defined, an error is thrown. If the desired
property is defined more than once, and is not consistently defined,
then an error is thrown. If the desired property is defined once, or is
defined consistently, then that value is returned.